+++ src/main/java/org/plumelib/util/EntryReader.java
@@ -60,7 +60,7 @@ import org.checkerframework.checker.regex.qual.Regex;
  *
  * <pre>{@code
  * // EntryReader constructor args are: filename, comment regexp, include regexp
- * try (EntryReader er = new EntryReader(filename, "^#.*", null)) {
+ * try (EntryReader er = new EntryReader(filename, false, "^#.*", null)) {
  *   for (String line : er) {
  *     ...
  *   }
@@ -108,6 +108,9 @@ public class EntryReader extends LineNumberReader implements Iterable<String>, I
    * @see #entryStartRegex
    */
   public @MonotonicNonNull Pattern entryStopRegex = null;
+
+  /** If true, then entries are separated by two blank lines rather than one. */
+  public boolean twoBlankLines = false;
 
   //
   // Internal implementation variables
@@ -196,7 +199,7 @@ public class EntryReader extends LineNumberReader implements Iterable<String>, I
 
     /**
      * Returns a substring of the entry body that matches the specified regular expression. If no
-     * match is found, returns the firstLine.
+     * match is found, returns {@link #firstLine}.
      *
      * @param re regex to match
      * @return a substring that matches re
@@ -228,6 +231,7 @@ public class EntryReader extends LineNumberReader implements Iterable<String>, I
    * @param in source from which to read entries
    * @param charsetName the character set to use
    * @param filename non-null file name for stream being read
+   * @param twoBlankLines true if entries are separated by two blank lines rather than one
    * @param commentRegexString regular expression that matches comments. Any text that matches
    *     commentRegex is removed. A line that is entirely a comment is ignored.
    * @param includeRegexString regular expression that matches include directives. The expression
@@ -239,10 +243,16 @@ public class EntryReader extends LineNumberReader implements Iterable<String>, I
       @MustCallAlias InputStream in,
       String charsetName,
       String filename,
+      boolean twoBlankLines,
       @Nullable @Regex String commentRegexString,
       @Nullable @Regex(1) String includeRegexString)
       throws UnsupportedEncodingException {
-    this(new InputStreamReader(in, charsetName), filename, commentRegexString, includeRegexString);
+    this(
+        new InputStreamReader(in, charsetName),
+        filename,
+        twoBlankLines,
+        commentRegexString,
+        includeRegexString);
   }
 
   /**
@@ -257,7 +267,7 @@ public class EntryReader extends LineNumberReader implements Iterable<String>, I
   public @MustCallAlias EntryReader(
       @MustCallAlias InputStream in, String charsetName, String filename)
       throws UnsupportedEncodingException {
-    this(in, charsetName, filename, null, null);
+    this(in, charsetName, filename, false, null, null);
   }
 
   // Inputstream (no charset) constructors
@@ -277,7 +287,7 @@ public class EntryReader extends LineNumberReader implements Iterable<String>, I
       String filename,
       @Nullable @Regex String commentRegexString,
       @Nullable @Regex(1) String includeRegexString) {
-    this(new InputStreamReader(in, UTF_8), filename, commentRegexString, includeRegexString);
+    this(new InputStreamReader(in, UTF_8), filename, false, commentRegexString, includeRegexString);
   }
 
   /**
@@ -286,7 +296,7 @@ public class EntryReader extends LineNumberReader implements Iterable<String>, I
    *
    * @param in the InputStream
    * @param filename the file name
-   * @see #EntryReader(InputStream,String,String,String,String)
+   * @see #EntryReader(InputStream,String,String,String)
    */
   public @MustCallAlias EntryReader(@MustCallAlias InputStream in, String filename) {
     this(in, filename, null, null);
@@ -374,6 +384,7 @@ public class EntryReader extends LineNumberReader implements Iterable<String>, I
    *
    * @param reader source from which to read entries
    * @param filename file name corresponding to reader, for use in error messages
+   * @param twoBlankLines true if entries are separated by two blank lines rather than one
    * @param commentRegexString regular expression that matches comments. Any text that matches
    *     commentRegex is removed. A line that is entirely a comment is ignored
    * @param includeRegexString regular expression that matches include directives. The expression
@@ -383,12 +394,14 @@ public class EntryReader extends LineNumberReader implements Iterable<String>, I
   public @MustCallAlias EntryReader(
       @MustCallAlias Reader reader,
       String filename,
+      boolean twoBlankLines,
       @Nullable @Regex String commentRegexString,
       @Nullable @Regex(1) String includeRegexString) {
     // we won't use superclass methods, but passing null as an argument
     // leads to a NullPointerException.
     super(DummyReader.it);
     readers.addFirst(new FlnReader(reader, filename));
+    this.twoBlankLines = twoBlankLines;
     if (commentRegexString == null) {
       commentRegex = null;
     } else {
@@ -405,10 +418,10 @@ public class EntryReader extends LineNumberReader implements Iterable<String>, I
    * Create an EntryReader that does not support comments or include directives.
    *
    * @param reader source from which to read entries
-   * @see #EntryReader(Reader,String,String,String)
+   * @see #EntryReader(Reader,String,boolean,String,String)
    */
   public @MustCallAlias EntryReader(@MustCallAlias Reader reader) {
-    this(reader, reader.toString(), null, null);
+    this(reader, reader.toString(), false, null, null);
   }
 
   // Path constructors
@@ -417,6 +430,7 @@ public class EntryReader extends LineNumberReader implements Iterable<String>, I
    * Create an EntryReader.
    *
    * @param path initial file to read
+   * @param twoBlankLines true if entries are separated by two blank lines rather than one
    * @param commentRegex regular expression that matches comments. Any text that matches
    *     commentRegex is removed. A line that is entirely a comment is ignored.
    * @param includeRegex regular expression that matches include directives. The expression should
@@ -424,9 +438,13 @@ public class EntryReader extends LineNumberReader implements Iterable<String>, I
    * @throws IOException if there is a problem reading the file
    */
   public EntryReader(
-      Path path, @Nullable @Regex String commentRegex, @Nullable @Regex(1) String includeRegex)
+      Path path,
+      boolean twoBlankLines,
+      @Nullable @Regex String commentRegex,
+      @Nullable @Regex(1) String includeRegex)
       throws IOException {
-    this(FilesPlume.newFileReader(path), path.toString(), commentRegex, includeRegex);
+    this(
+        FilesPlume.newFileReader(path), path.toString(), twoBlankLines, commentRegex, includeRegex);
   }
 
   /**
@@ -434,10 +452,10 @@ public class EntryReader extends LineNumberReader implements Iterable<String>, I
    *
    * @param path the file to read
    * @throws IOException if there is a problem reading the file
-   * @see #EntryReader(File,String,String)
+   * @see #EntryReader(File,boolean,String,String)
    */
   public EntryReader(Path path) throws IOException {
-    this(path, null, null);
+    this(path, false, null, null);
   }
 
   /**
@@ -446,10 +464,10 @@ public class EntryReader extends LineNumberReader implements Iterable<String>, I
    * @param path the file to read
    * @param charsetName the character set to use
    * @throws IOException if there is a problem reading the file
-   * @see #EntryReader(Path,String,String)
+   * @see #EntryReader(Stream,String,String,boolean,String,String)
    */
   public EntryReader(Path path, String charsetName) throws IOException {
-    this(FilesPlume.newFileInputStream(path), charsetName, path.toString(), null, null);
+    this(FilesPlume.newFileInputStream(path), charsetName, path.toString(), false, null, null);
   }
 
   // File constructors
@@ -458,6 +476,7 @@ public class EntryReader extends LineNumberReader implements Iterable<String>, I
    * Create an EntryReader.
    *
    * @param file initial file to read
+   * @param twoBlankLines true if entries are separated by two blank lines rather than one
    * @param commentRegex regular expression that matches comments. Any text that matches
    *     commentRegex is removed. A line that is entirely a comment is ignored.
    * @param includeRegex regular expression that matches include directives. The expression should
@@ -465,9 +484,13 @@ public class EntryReader extends LineNumberReader implements Iterable<String>, I
    * @throws IOException if there is a problem reading the file
    */
   public EntryReader(
-      File file, @Nullable @Regex String commentRegex, @Nullable @Regex(1) String includeRegex)
+      File file,
+      boolean twoBlankLines,
+      @Nullable @Regex String commentRegex,
+      @Nullable @Regex(1) String includeRegex)
       throws IOException {
-    this(FilesPlume.newFileReader(file), file.toString(), commentRegex, includeRegex);
+    this(
+        FilesPlume.newFileReader(file), file.toString(), twoBlankLines, commentRegex, includeRegex);
   }
 
   /**
@@ -475,10 +498,10 @@ public class EntryReader extends LineNumberReader implements Iterable<String>, I
    *
    * @param file the file to read
    * @throws IOException if there is a problem reading the file
-   * @see #EntryReader(File,String,String)
+   * @see #EntryReader(File,boolean,String,String)
    */
   public EntryReader(File file) throws IOException {
-    this(file, null, null);
+    this(file, false, null, null);
   }
 
   /**
@@ -487,10 +510,10 @@ public class EntryReader extends LineNumberReader implements Iterable<String>, I
    * @param file the file to read
    * @param charsetName the character set to use
    * @throws IOException if there is a problem reading the file
-   * @see #EntryReader(File,String,String)
+   * @see #EntryReader(File,boolean,String,String)
    */
   public EntryReader(File file, String charsetName) throws IOException {
-    this(FilesPlume.newFileInputStream(file), charsetName, file.toString(), null, null);
+    this(FilesPlume.newFileInputStream(file), charsetName, file.toString(), false, null, null);
   }
 
   // Filename constructors
@@ -499,19 +522,21 @@ public class EntryReader extends LineNumberReader implements Iterable<String>, I
    * Create a new EntryReader starting with the specified file.
    *
    * @param filename initial file to read
+   * @param twoBlankLines true if entries are separated by two blank lines rather than one
    * @param commentRegex regular expression that matches comments. Any text that matches {@code
    *     commentRegex} is removed. A line that is entirely a comment is ignored.
    * @param includeRegex regular expression that matches include directives. The expression should
    *     define one group that contains the include file name.
    * @throws IOException if there is a problem reading the file
-   * @see #EntryReader(File,String,String)
+   * @see #EntryReader(File,boolean,String,String)
    */
   public EntryReader(
       String filename,
+      boolean twoBlankLines,
       @Nullable @Regex String commentRegex,
       @Nullable @Regex(1) String includeRegex)
       throws IOException {
-    this(new File(filename), commentRegex, includeRegex);
+    this(new File(filename), twoBlankLines, commentRegex, includeRegex);
   }
 
   /**
@@ -519,10 +544,10 @@ public class EntryReader extends LineNumberReader implements Iterable<String>, I
    *
    * @param filename source from which to read entries
    * @throws IOException if there is a problem reading the file
-   * @see #EntryReader(String,String,String)
+   * @see #EntryReader(String,boolean,String,String)
    */
   public EntryReader(String filename) throws IOException {
-    this(filename, null, null);
+    this(filename, false, null, null);
   }
 
   /**
@@ -531,10 +556,10 @@ public class EntryReader extends LineNumberReader implements Iterable<String>, I
    * @param filename source from which to read entries
    * @param charsetName the character set to use
    * @throws IOException if there is a problem reading the file
-   * @see #EntryReader(String,String,String)
+   * @see #EntryReader(String,boolean,String,String)
    */
   public EntryReader(String filename, String charsetName) throws IOException {
-    this(Files.newInputStream(Path.of(filename)), charsetName, filename, null, null);
+        this(Files.newInputStream(Path.of(filename)), charsetName, filename, false, null, null);
   }
 
   //
@@ -542,10 +567,12 @@ public class EntryReader extends LineNumberReader implements Iterable<String>, I
   //
 
   /**
-   * Read a line, ignoring comments and processing includes. Note that a line that is completely a
-   * comment is completely ignored (and not returned as a blank line). Returns null at end of file.
+   * Read a line, ignoring comments and processing includes. Returns null at end of file.
    *
-   * @return the string that was read, or null at end of file
+   * <p>A line that is completely a comment is ignored (and not returned as a blank line).
+   *
+   * @return the string that was read, not including any line termination characters, or null at end
+   *     of file
    */
   @Override
   public @Nullable String readLine(@GuardSatisfied EntryReader this) throws IOException {
@@ -684,9 +711,10 @@ public class EntryReader extends LineNumberReader implements Iterable<String>, I
   }
 
   /**
-   * Returns the next entry (paragraph) in the file. Entries are separated by blank lines unless the
-   * entry started with {@link #entryStartRegex} (see {@link #setEntryStartStop}). If no more
-   * entries are available, returns null.
+   * Returns the next entry (paragraph) in the file. If no more entries are available, returns null.
+   *
+   * <p>Entries are separated by one or two blank lines (two, if {@link #twoBlankLines} is true),
+   * unless the entry started with {@link #entryStartRegex} (see {@link #setEntryStartStop}).
    *
    * @return the next entry (paragraph) in the file
    * @throws IOException if there is a problem reading the file
@@ -702,7 +730,7 @@ public class EntryReader extends LineNumberReader implements Iterable<String>, I
       return null;
     }
 
-    StringBuilder body = new StringBuilder(1000);
+    StringBuilder body = new StringBuilder(500);
     String filename = getFileName();
     long lineNumber = getLineNumber();
 
@@ -756,15 +784,36 @@ public class EntryReader extends LineNumberReader implements Iterable<String>, I
 
       String description = line;
 
-      // Read until we find another blank line.
-      while (line != null && !line.isBlank() && filename.equals(getFileName())) {
+      // Read until we find blank line(s) that separate entries.
+      String blankLineFound = null;
+      while ((line != null) && filename.equals(getFileName())) {
+        if (line.isBlank()) {
+          if (!twoBlankLines) {
+            break;
+          } else if (blankLineFound != null) {
+            break;
+          } else {
+            blankLineFound = line;
+            line = readLine();
+            continue;
+          }
+        }
+
+        // The line is not blank.
+
+        if (blankLineFound != null) {
+          body.append(blankLineFound);
+          body.append(lineSep);
+          blankLineFound = null;
+        }
+
         body.append(line);
         body.append(lineSep);
         line = readLine();
       }
 
       // If this entry was terminated by the start of a new input file
-      // put that line back
+      // put that line back.
       if ((line != null) && !filename.equals(getFileName())) {
         putback(line);
       }
@@ -925,9 +974,9 @@ public class EntryReader extends LineNumberReader implements Iterable<String>, I
           "EntryReader sample program requires 1-3 args: filename [commentRegex [includeRegex]]");
       System.exit(1);
     }
-    String filename = args[0];
-    String commentRegex = null;
-    String includeRegex = null;
+    final String filename = args[0];
+
+    final String commentRegex;
     if (args.length >= 2) {
       commentRegex = args[1];
       if (!RegexUtil.isRegex(commentRegex)) {
@@ -938,9 +987,15 @@ public class EntryReader extends LineNumberReader implements Iterable<String>, I
                 + RegexUtil.regexError(commentRegex));
         System.exit(1);
       }
+    } else {
+      commentRegex = null;
     }
+
+    final @Regex(1) String includeRegex;
     if (args.length >= 3) {
-      includeRegex = args[2];
+      @SuppressWarnings("regex:assignment") // about to be checked; flow isn't properly refining?
+      @Regex(1) String arg3 = args[2];
+      includeRegex = arg3;
       if (!RegexUtil.isRegex(includeRegex, 1)) {
         System.err.println(
             "Error parsing include regex \""
@@ -948,10 +1003,13 @@ public class EntryReader extends LineNumberReader implements Iterable<String>, I
                 + "\": "
                 + RegexUtil.regexError(includeRegex));
         System.exit(1);
+        throw new Error("unreachable");
       }
+    } else {
+      includeRegex = null;
     }
-    try (EntryReader reader = new EntryReader(filename, commentRegex, includeRegex)) {
 
+    try (EntryReader reader = new EntryReader(filename, false, commentRegex, includeRegex)) {
       String line = reader.readLine();
       while (line != null) {
         System.out.printf("%s: %d: %s%n", reader.getFileName(), reader.getLineNumber(), line);